Toggle cells on the tick
As I'm figuring out how React works, I'm trying to add things incrementally.
Before following Conway's rules to determine who should be alive and who should
be dead on each click, I'm just toggling one cell each time.

I've realized that this isn't going to work quite like I wanted it to.  While
clicking a cell right now toggles it's living/dead status correctly, this change
is not reflected in the Game Boards' cells, meaning that the change is undone on
the next tick.  This marks a major difference between React and, for example,
Ember.js, which would bind those changes both ways to that the change in the
state of a Cell component would be reflected in the change to the "model"beneath
the component, and therefor reflected in the Rows structure in the Game Board.
I think, based on seeing some example code online, that I'll have to handle the
click event through the Game Board, which I really don't like.  That, or I need
to start looking into two-way binding in React.