Maintain cell state when toggling through a click
Now here's something I really don't like about React.  State is bound from a
parent to the child, not the other way around (at least, not by default).
Because of that, I had to create the toggle method on the Game Board, and pass
that method from the Board to the Row to the Cell, so that the Cell triggers the
method on the Game board instead of a method that it has locally.

I also learned that you can set jump points in JavaScript, which is... weird.  I
don't know why I assumed that JS didn't have them, but it allowed me to jump out
of a nested `for` loop prematurely, so... Whatever works, right?